1) Move AsyncLazy classes to separate files 2) Fix test failure 3) Fix formatting request 4) Change AsyncLazy ctors to protected 5) Rename ResetComputeFunctions to ClearComputeFunctions
